Rename contrib.captcha to contrib.waf
Rename the contrib.captcha package to contrib.waf to reflect these integrations as AWS WAF JavaScript challenge solvers (distinct from legacy OCR image captchas). Move/rename modules (anticaptcha, capsolver, twocaptcha, base), update imports, and adjust tests to import from amazonorders.contrib.waf. Update docs, index, troubleshooting, README, and CHANGELOG to reference the new waf pages and examples (docs/captcha.rst -> docs/waf.rst) and update auth_forms_classes examples and links.
